Refactor dictionary and sensitive mongo field encryption by moving it to server_utils/encryption
