+++ monkey/monkey_island/cc/models/report/__init__.py
@@ -1,0 +1,1 @@
+from .report_dal import save_report, get_report
