+++ monkey/monkey_island/cc/server_utils/encryption/dict_encryption/field_encryptors/string_list_encryptor.py
@@ -1,7 +1,9 @@
 from typing import List
 
 from monkey_island.cc.server_utils.encryption import get_datastore_encryptor
-from monkey_island.cc.utils.field_encryptors import IFieldEncryptor
+from monkey_island.cc.server_utils.encryption.dict_encryption.field_encryptors import (
+    IFieldEncryptor,
+)
 
 
 class StringListEncryptor(IFieldEncryptor):
