--- monkey/monkey_island/cc/utils/__init__.py
@@ -1,1 +1,0 @@
-from .dict_encryptor import FieldNotFoundError, SensitiveField
