+++ monkey/tests/unit_tests/monkey_island/cc/models/utils/field_encryptors/test_string_list_encryptor.py
@@ -1,7 +1,9 @@
 import pytest
 
 from monkey_island.cc.server_utils.encryption import initialize_datastore_encryptor
-from monkey_island.cc.utils.field_encryptors import StringListEncryptor
+from monkey_island.cc.server_utils.encryption.dict_encryption.field_encryptors import (
+    StringListEncryptor,
+)
 
 MOCK_STRING_LIST = ["test_1", "test_2"]
 EMPTY_LIST = []
